refactor: move availability debug time parser
