+++ src/pages/availability/week/AvailabilityWeekPage.tsx
@@ -44,29 +44,7 @@ import {
 	WeekTitleBlock,
 } from './AvailabilityWeekPage.styles';
 import type { IWeekSlot } from './AvailabilityWeekPage.types';
-
-const parseDebugNowMinutes = (value: string | null): number | null => {
-	if (!value || !import.meta.env.DEV) return null;
-
-	const match = value.match(/^(\d{1,2}):(\d{2})$/);
-	if (!match) return null;
-
-	const hours = Number(match[1]);
-	const minutes = Number(match[2]);
-
-	if (
-		Number.isNaN(hours) ||
-		Number.isNaN(minutes) ||
-		hours < 0 ||
-		hours > 23 ||
-		minutes < 0 ||
-		minutes > 59
-	) {
-		return null;
-	}
-
-	return hours * 60 + minutes;
-};
+import { parseDebugNowMinutes } from './AvailabilityWeekPage.utils';
 
 export const AvailabilityWeekPage = () => {
 	const todayCardId = 'availability-week-mobile-today';
